Adding comments and fixing user interface
The key part of this is the fixes that were generated after the fist hub
test was completed. This includes removing the continue button, giving
more user feed back and making is to that when multiple statused
daysimeters are on the hub it will only let you stop and download them.
Comments were also added to a majority of the longer hub functions. More
will be added as functions settle down.